Sketching out draft for running arbitrary scripts;
So there's a lot going on in this PR. But I am giving users the ability
to set four pairs of options which consist of a required working script
and an optional icon. I then use these variables to check if the
executable is runnable and then create a half-pill design like the
others.