Initial support for window icons;
So Tmux has a default value to display window icons. This patch adds
that functionality and more. There's also the Nerd Font support for
window icons. There are defaults too. But I do need to make some
documentation here and also tweak the code to make it clearer what and
how things are configured.